Widen the refcount, because 65536 copies of one packet freed it
A clean-room review pass found a use-after-free reachable through
nothing but the public API. Demonstrated before fixing, under ASan:

    auto original = rx.pop_packet();
    std::vector<Ref> copies;             // 65536 copies of it
    copies.pop_back();                   // destroy exactly ONE
    original.data();                     // heap-use-after-free

PacketRef's copy constructor does ++refs unchecked, so a uint16_t
counter wraps: (1 + 65536) mod 65536 is 1 again, and dropping any single
handle takes it to zero while 65536 others are still alive. ASan named it
exactly, with the free coming from PacketRef::release() and the read from
RxPacket::data().

refs becomes uint32_t. Saturating instead would leak by design and hide
the mistake; a wider counter puts the wrap beyond what handles to one
frame can reach, since 2^32 live PacketRefs would need 32 GiB of them.
The cost is measured: sizeof(RxPacket) goes 12 -> 16 on ARM32 and stays
24 on x86-64 where the padding already existed — for the bench's 4-block
pool, 16 bytes of bss. The regression test copies 70000 times, past where
the old width broke and nowhere near where this one does.

Two contract defects went with it.

TxAllocation.h still documented the pre-prefix obligation, "the block
holds at least cobs_max_wire_size(capacity) bytes". Both shipped policies
and the contract test have been header-inclusive since the length prefix
landed, but this header is where somebody writing their own policy will
read the rule from — so it was not a stale comment, it was a set of
instructions for building an allocator that CobsMsg::encode() overruns by
one or two bytes. It now states tx_storage_size_for_capacity() and says
why open-coding it is the trap.

COBS_POOL_CHECKS defaulted to 0 under NDEBUG, so the shipped
configuration had weaker safety semantics than the one every test ran
against: the comments and the suite both promise that a double free is
counted and ignored, while a release build would have threaded the same
block onto the free list twice. It now defaults to on in every build and
is an explicit opt-out — the cost is a walk of a 2-to-8-entry free list
on a path that is never in an ISR. And the suite now builds
test_block_pool and test_allocators a second time with -DNDEBUG, so the
release path is tested rather than assumed.

New boundary battery at the edges of the length field itself: bodies of
255 (the last a one-byte header can describe), 256, and 65535. The last
is the one worth having — a 65535-byte body fits RxPacket::size exactly
while its 65537-byte decoded frame does not, so anything casting the
decoded size rather than the body would truncate to 1 and deliver a
one-byte packet. CobsRx subtracts the header first; now a test says so.

Stale comments: CobsRx explained that max_receive_size was the bad name
(it was max_decoded_size), RxPacket still said "three friends" after the
allocator stopped being one, and test_block_pool still called 01 00 a
minimal TX frame.

Verification: 22987 checks, 0 failures, clean under ASan+UBSan in WSL.
Nine binaries now — the seven suites plus two rebuilt with -DNDEBUG:

    test_decoder 19981, test_block_pool 16, test_allocators 85,
    test_cobs_rx 137, test_encoder 2370, test_cobs_msg 140,
    test_cobs 157, and 16 + 85 again under -DNDEBUG

The ASan probe that reproduced the bug now runs clean. Cortex-M7 at -Os:
3016 bytes of text (+6) and 1824 of bss (+16), no malloc, no operator
new.

+++ cobs/RxPacket.h
@@ -23,7 +23,7 @@
  * — the exact counterpart of the writable_payload() hole, and closed the same
  * way. `refs`, `next_ready` and `owner` are the ownership bookkeeping the
  * whole lifetime model rests on; code that can write them can free a live
- * packet or leak a dead one. Three friends can, and each has to:
+ * packet or leak a dead one. Two friends can, and each has to:
  *
  *     CobsRx      builds the packet, stamps `owner`, and threads it onto
  *                 the ready queue
@@ -67,7 +67,26 @@ struct RxPacket final {
 	}
 
 private:
-	uint16_t  refs       = 1;       // the creator holds the first reference
+	/*
+	 * 32 bits, not 16, and that is a bug fix rather than a preference.
+	 *
+	 * PacketRef's copy constructor does `++refs` with no check, so a 16-bit
+	 * counter wraps after 65536 copies — reachable through nothing but the
+	 * public API, and fatal:
+	 *
+	 *     65536 copies      stored refs = (1 + 65536) mod 65536 = 1
+	 *     destroy ONE       stored refs 1 -> 0, the packet is freed
+	 *     the other 65536 handles are still alive
+	 *
+	 * ASan calls it a heap-use-after-free, correctly. Saturating instead would
+	 * leak by design and hide the mistake; a wider counter simply puts the
+	 * wrap beyond anything a single frame's handles can reach, since 2^32
+	 * live PacketRefs would need 32 GiB of them.
+	 *
+	 * The cost is four bytes per RX block on a 32-bit target (12 -> 16), and
+	 * nothing at all on x86-64, where the padding already existed.
+	 */
+	uint32_t  refs       = 1;       // the creator holds the first reference
 	uint16_t  size       = 0;       // decoded bytes; set on FrameComplete
 	RxPacket* next_ready = nullptr; // intrusive ready-queue link (§6.2)
 	Allocator* owner     = nullptr; // who reclaims this block
